finish updating test suite and update demo
