+++ code/demo.py
@@ -20,11 +20,11 @@ import matplotlib.pyplot as plt
 # Demo of chemkin module
 # Parses rxns.xml and calculates chemical reaction rates for each of the six species
 
-print(ck.get_reactions("demo_xmls/rxns.xml"))
-
 # Pull data from .xml file, and set concentration rates
-demo_reaction = ck.Reaction(ck.get_reactions("demo_xmls/rxns.xml"))
+demo_reaction = ck.Reaction("demo_xmls/rxns.xml")
 concs = np.array([2.0, 1.0, 2.0, 0.5, 1.5, 2.5]).reshape(-1, 1)
+
+print(demo_reaction.get_params())
 
 # Calculate reaction rates vs T
 T_range = np.arange(1273,5273)
