+++ code/test_xmls/reaction_rate_1.xml
@@ -1,0 +1,38 @@
+<?xml version="1.0"?>
+
+<ctml>
+
+    <phase>
+        <speciesArray> H2 O OH </speciesArray>
+    </phase>
+
+    <reactionData id="test_mechanism">
+
+        <!-- reaction 01  -->
+        <reaction reversible="no" type="Elementary" id="reaction01">
+            <equation>H2 + 2O =] 2OH</equation>
+            <rateCoeff>
+                <Constant>
+                    <k>10</k>
+                </Constant>
+            </rateCoeff>
+            <reactants>H2:1 O:2</reactants>
+            <products>OH:2</products>
+        </reaction>
+
+        <!-- reaction 02 -->
+        <reaction reversible="no" type="Elementary" id="reaction02">
+            <equation>2H2 + 2OH =] O + OH </equation>
+            <rateCoeff>
+                  <Constant>
+                      <k>10</k>
+                  </Constant>
+            </rateCoeff>
+            <reactants>H2:2 OH:2</reactants>
+            <products>O:1 OH:1</products>
+        </reaction>
+        
+
+    </reactionData>
+
+</ctml>
